+++ clang-tools-extra/test/clang-tidy/checkers/autosar/Inputs/Std_Types.h
@@ -1,0 +1,22 @@
+#ifndef STD_TYPES_H
+#define STD_TYPES_H
+
+#include "Platform_Types.h"
+
+#define STD_OFF 0x00u
+#define STD_ON 0x01u
+
+#define STD_LOW 0x00u
+#define STD_HIGH 0x01u
+
+#define E_OK 0u
+#define E_NOT_OK 1u
+
+#define RTE_E_OK 0u
+#define RTE_E_INVALID 1u
+
+#define NULL_PTR ((void *) 0)
+
+typedef uint8 Std_ReturnType;
+
+#endif /* STD_TYPES_H */