+++ clang-tools-extra/test/clang-tidy/checkers/autosar/Inputs/Swc_MemMap_Extended.h
@@ -1,0 +1,1 @@
+/* Stubbed memory mapping header */