+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -176,6 +176,12 @@ New checks
   since C23 and C++23, uses of ``#elif`` that can be simplified to ``#elifdef``
   or ``#elifndef``.
 
+- New :doc:`autosar-memory-mapping
+  <clang-tidy/checks/autosar/memory-mapping>` check.
+
+  Detects multiple potential issues with regards to the AUTOSAR memory mapping
+  specification including unmapped or wrongly mapped variables and functions.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^